Add pure Lua sentence and syllable parsers
Previously I was just using the `style` command to calculate the SMOG
index for me and show that to the user. I always had the intention to
write a complete syllable and sentence parser in Lua so anyone could run
this. After a lot of procrastinating I finally got around to doing it.