fix: No more nasty viml or bugs
The VimL in the plugin/readability.vim has been eradicated and
completely replaced with Lua. The two commands to calculate the
Flesch-Kincaid and SMOG grade levels now work correctly. The layout of
the files has also been changed around a bit. I also changed the en.txt
file to a Lua file that simply returns a table to skip having to read
explicitly read the text file and split it into separate lines.
Everything is working as it's supposed to!